+++ backend/libraries/group_chat_core/src/lib.rs
@@ -67,6 +67,39 @@ pub struct GroupChatCore {
 
 #[allow(clippy::too_many_arguments)]
 impl GroupChatCore {
+    pub fn dedupe_at_everyone_mentions(&mut self) {
+        let mut at_everyone_mentions = BTreeMap::new();
+
+        for member in self.members.iter_mut() {
+            member.mentions.remove_at_everyone_mentions(&at_everyone_mentions);
+
+            let mut updated = false;
+            for mention in member.mentions.iter_potential_at_everyone_mentions() {
+                if let Some(event_wrapper) =
+                    self.events
+                        .event_wrapper_internal(EventIndex::default(), None, mention.message_index.into())
+                {
+                    if let Some(message) = event_wrapper.event.into_message() {
+                        if is_everyone_mentioned(&message.content) {
+                            at_everyone_mentions.insert(
+                                mention.message_index,
+                                (
+                                    event_wrapper.timestamp,
+                                    AtEveryoneMention::new(message.sender, message.message_id, message.message_index),
+                                ),
+                            );
+                            updated = true;
+                        }
+                    }
+                }
+            }
+
+            if updated {
+                member.mentions.remove_at_everyone_mentions(&at_everyone_mentions);
+            }
+        }
+    }
+
     pub fn new(
         chat: MultiUserChat,
         created_by: UserId,
@@ -2275,7 +2308,7 @@ pub enum MinVisibleEventIndexResult {
 }
 
 #[derive(Serialize, Deserialize)]
-struct AtEveryoneMention(UserId, MessageId, MessageIndex);
+pub struct AtEveryoneMention(UserId, MessageId, MessageIndex);
 
 impl AtEveryoneMention {
     fn new(sender: UserId, message_id: MessageId, message_index: MessageIndex) -> AtEveryoneMention {